requirements.txt
flask==2.3.2
flask-restful==0.3.10
flask-sqlalchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
python-dotenv==1.0.0
flask-jwt-extended==4.5.2
apispec==6.3.0